feat: add ability to ignore specific keys (#20)
* feat: add ability to ignore specfic CloudFormation keys

The idea of this feature is to minimize the diff size by removing irrelevant portions of the CloudFormation template. A good example of this is if your project contains a large number of lambdas whose S3Key changes on every synth. If utilizing `ignore-keys`, the S3Key attribute is removed from the diff and only more relevant changes will be commented to the PR.

* chore: bumping version

* chore: rename IGNORE_KEYS to CDK_DIFF_IGNORE_KEYS

Co-authored-by: Mark Nielsen <mark.nielsen@blackboard.com>

* fix: missed renaming from IGNORE_KEYS to CDK_DIFF_IGNORE_KEYS

---------

Co-authored-by: Mark Nielsen <mark.nielsen@blackboard.com>